git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/aims/aims-free/trunk@41917 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ aimsalgo/src/aimsalgo/geodesicpath/geodesic_mesh.h
@@ -38,7 +38,7 @@ public:
   template<class Points, class Faces>
   void initialize_mesh_data(Points& p, Faces& tri, const float* curvature, int mode,int strain);    //build mesh from regular point-triangle representation
 
-  void update_weight(const float *curvature,int mode,float strain, float sigmo);    //build internal structure of the mesh
+  void update_weight(const float *curvature,int mode,double strain, double sigmo);    //build internal structure of the mesh
 
   std::vector<Vertex>& vertices(){return m_vertices;};
   std::vector<Edge>& edges(){return m_edges;};
@@ -416,7 +416,7 @@ inline void Mesh::build_adjacencies(const float *curvature, int mode, int strain
   //assert(verify());
 }
 
-inline void Mesh::update_weight(const float *curvature, int mode,float strain,float sigmo)
+inline void Mesh::update_weight(const float *curvature, int mode,double strain,double sigmo)
 {
   //    Vertex->adjacent Faces
   std::vector<unsigned> count(m_vertices.size()); //count adjacent vertices
@@ -519,15 +519,15 @@ inline void Mesh::update_weight(const float *curvature, int mode,float strain,fl
       // Sulcal
       if (mode == 1)
       {
-        a1 = pow ((1.0)/(1.0 + exp(-strain*curvature[v1->id()])), sigmo);
-        a2 = pow ((1.0)/(1.0 + exp(-strain*curvature[v2->id()])), sigmo);
+        a1 = pow ((double)(1.0)/(1.0 + exp(-strain*curvature[v1->id()])), (double)sigmo);
+        a2 = pow ((double)(1.0)/(1.0 + exp(-strain*curvature[v2->id()])), (double)sigmo);
       }
 
       // Gyral
       if (mode == 2)
       {
-        a1 = pow ((1.0)/(1.0 + exp(strain*curvature[v1->id()])), sigmo);
-        a2 = pow ((1.0)/(1.0 + exp(strain*curvature[v2->id()])), sigmo);
+        a1 = pow ((double)(1.0)/(1.0 + exp(strain*curvature[v1->id()])), (double)sigmo);
+        a2 = pow ((double)(1.0)/(1.0 + exp(strain*curvature[v2->id()])), (double)sigmo);
       }
 
       e.length() = (v1->distance(v2) * (a1 + a2));
